Ensure mysql service is started after install
On some older versions of Ubuntu (e.g. 20.04), the mysql service is not
automatically started after installation. This can cause configuration
to fail. In this change, we add `service mysql start` right before
configuring the mysql databases.

We use `service` instead of `systemctl` in order to account for other
VM hosts such as Docker, which uses init instead of systemd. On systems
with systemd, `service mysql start` will be aliased appropriately.